Сделал проверку на вклады

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('password');
             $table->string('name');
             $table->string('surname');
-            $table->integer('balance')->default(0);
+            $table->float('balance')->default(0.00);
             $table->string('role')->default(0);
             $table->rememberToken();
             $table->timestamps();
